requirements.txt
flask
tensorflow==2.13.0
nltk
numpy
scikit-learn
pyyaml
sentence-transformers
faiss-cpu
requests
gunicorn